+++ lib/ton_sdk_client/boc.rb
@@ -1,5 +1,10 @@
 module TonSdk
   module Boc
+
+    #
+    # types
+    #
+
     class ParamsOfParse
       attr_reader :boc
 
@@ -62,9 +67,35 @@ module TonSdk
       end
     end
 
-    def self.parse_message(ctx, pr1)
-      pr_json = pr1.to_h.to_json
-      Interop::request_to_native_lib(ctx, "boc.parse_message", pr_json) do |resp|
+    class ParamsOfGetBocHash
+      attr_reader :boc
+
+      def initialize(a)
+        @boc = a
+      end
+
+      def to_h
+        {
+          boc: @boc
+        }
+      end
+    end
+
+    class ResultOfGetBlockchainConfig
+      attr_reader :hash
+
+      def initialize(a)
+        @hash = a
+      end
+    end
+
+
+    #
+    # functions
+    #
+
+    def self.parse_message(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.parse_message", pr_s.to_h.to_json) do |resp|
         if resp.success?
           yield NativeLibResponsetResult.new(
             result: ResultOfParse.new(resp.result["parsed"])
@@ -75,9 +106,8 @@ module TonSdk
       end
     end
 
-    def self.parse_transaction(ctx, pr1)
-      pr_json = pr1.to_h.to_json
-      Interop::request_to_native_lib(ctx, "boc.parse_transaction", pr_json) do |resp|
+    def self.parse_transaction(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.parse_transaction", pr_s.to_h.to_json) do |resp|
         if resp.success?
           yield NativeLibResponsetResult.new(
             result: ResultOfParse.new(resp.result["parsed"])
@@ -88,9 +118,8 @@ module TonSdk
       end
     end
 
-    def self.parse_account(ctx, pr1)
-      pr_json = pr1.to_h.to_json
-      Interop::request_to_native_lib(ctx, "boc.parse_account", pr_json) do |resp|
+    def self.parse_account(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.parse_account", pr_s.to_h.to_json) do |resp|
         if resp.success?
           yield NativeLibResponsetResult.new(
             result: ResultOfParse.new(resp.result["parsed"])
@@ -101,9 +130,8 @@ module TonSdk
       end
     end
 
-    def self.parse_block(ctx, pr1)
-      pr_json = pr1.to_h.to_json
-      Interop::request_to_native_lib(ctx, "boc.parse_block", pr_json) do |resp|
+    def self.parse_block(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.parse_block", pr_s.to_h.to_json) do |resp|
         if resp.success?
           yield NativeLibResponsetResult.new(
             result: ResultOfParse.new(resp.result["parsed"])
@@ -114,9 +142,8 @@ module TonSdk
       end
     end
 
-    def self.get_blockchain_config(ctx, pr1)
-      pr_json = pr1.to_h.to_json
-      Interop::request_to_native_lib(ctx, "boc.get_blockchain_config", pr_json) do |resp|
+    def self.get_blockchain_config(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.get_blockchain_config", pr_s.to_h.to_json) do |resp|
         if resp.success?
           yield NativeLibResponsetResult.new(
             result: ResultOfGetBlockchainConfig.new(resp.result["config_boc"])
@@ -127,12 +154,23 @@ module TonSdk
       end
     end
 
-    def self.parse_shardstate(ctx, pr1)
-      pr_json = pr1.to_h.to_json
-      Interop::request_to_native_lib(ctx, "boc.parse_shardstate", pr_json) do |resp|
+    def self.parse_shardstate(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.parse_shardstate", pr_s.to_h.to_json) do |resp|
         if resp.success?
           yield NativeLibResponsetResult.new(
             result: ResultOfParse.new(resp.result["parsed"])
+          )
+        else
+          yield resp
+        end
+      end
+    end
+
+    def self.get_boc_hash(ctx, pr_s)
+      Interop::request_to_native_lib(ctx, "boc.get_boc_hash", pr_s.to_h.to_json) do |resp|
+        if resp.success?
+          yield NativeLibResponsetResult.new(
+            result: ResultOfGetBocHash.new(resp.result["hash"])
           )
         else
           yield resp