+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+1.1.x
+-----
+* `Client`: `resolve_app_request()`
+* `Net`: `query()`, `suspend()`, `resume()`
+* New module `Debot`
+* `Boc`: New method `get_boc_hash()`
+
 1.0.0
 -----
 * The first version has been released.